update from origin

+++ build/build_release.sh
@@ -270,7 +270,9 @@ find src -name pom.xml -exec sed -i "s/$old_ver/$tag/g" {} \;
 find doc -name conf.py -exec sed -i "s/$old_ver/$tag/g" {} \;
 
 pushd src/release > /dev/null
-sed -i "s/$old_ver/$tag/g" *.xml installer/win/*.nsi installer/win/*.conf 
+shopt -s extglob
+sed -i "s/$old_ver/$tag/g" !(pom).xml installer/win/*.nsi installer/win/*.conf 
+shopt -u extglob
 popd > /dev/null
 
 pushd src > /dev/null
@@ -283,22 +285,49 @@ if [ -z $SKIP_BUILD ]; then
   # build the javadocs
   mvn javadoc:aggregate
 
-  # build the user docs
-  pushd ../doc/en/user > /dev/null
-  make clean html
-  make latex
-  cd build/latex
-  sed  "s/includegraphics/includegraphics[scale=0.5]/g" GeoServerUserManual.tex > manual.tex
-  # run pdflatex twice in a row to get the TOC, and ignore errors 
-  set +e
-  pdflatex -interaction batchmode manual.tex
-  pdflatex -interaction batchmode manual.tex
-  set -e
+  ##################
+  # Build the docs
+  ##################
 
-  cd ../../../developer
-  make clean html
+
+
+  pushd ../doc/en > /dev/null
+
+  # 2.11 and older uses make
+  if [ -e user/Makefile ]
+  then
+    # build the user docs
+    cd user
+    make clean html
+    make latex
+    cd build/latex
+
+    sed  "s/includegraphics/includegraphics[scale=0.5]/g" GeoServerUserManual.tex > manual.tex
+    # run pdflatex twice in a row to get the TOC, and ignore errors 
+    set +e
+    pdflatex -interaction batchmode manual.tex
+    pdflatex -interaction batchmode manual.tex
+    set -e
+
+    if [ ! -f manual.pdf ]; then
+      echo "Failed to build pdf manual. Printing latex log:"
+      cat manual.log
+    fi
+
+    # build the developer docs
+    cd ../../../developer
+    make clean html
+
+  # 2.12 and newer uses ant to do everything
+  else
+    ant clean user
+    ant user-pdf
+    ant developer
+  fi
 
   popd > /dev/null
+
+
 fi
 
 mvn $MAVEN_FLAGS assembly:attached
@@ -333,45 +362,37 @@ if [ -e developer ]; then
   unlink developer
 fi
 
-ln -sf ../../../doc/en/user/build/html user
-ln -sf ../../../doc/en/developer/build/html developer
+# paths for 2.12 and newer docbuild
+usertarget=target/user
+devtarget=target/developer
+# paths for 2.11 and older docbuild
+if [ -e ../../../doc/en/user/Makefile ]; then
+  usertarget=user/build
+  devtarget=developer/build
+fi
+
+ln -sf ../../../doc/en/$usertarget/html user
+ln -sf ../../../doc/en/$devtarget/html developer
+ln -sf ../../../doc/en/release/README.txt readme
+
 htmldoc=geoserver-$tag-htmldoc.zip
 if [ -e $htmldoc ]; then
   rm -f $htmldoc 
 fi
-zip -r $htmldoc user developer
+zip -r $htmldoc user developer readme
 unlink user
 unlink developer
+unlink readme
 
-# clean up source artifact
-if [ -e tmp ]; then
-  rm -rf tmp
-fi
-mkdir tmp
-src=geoserver-$tag-src.zip
-unzip -d tmp $src
-pushd tmp > /dev/null
-
-set +e && find . -type d -name target -exec rm -rf {} \; && set -e
-rm ../$src
-zip -r ../$src *
-
-popd > /dev/null
 popd > /dev/null
 
 echo "copying artifacts to $dist"
-cp $artifacts/../../../doc/en/user/build/latex/manual.pdf $dist/geoserver-$tag-user-manual.pdf
 cp $artifacts/*-plugin.zip $dist/plugins
 for a in `ls $artifacts/*.zip | grep -v plugin`; do
   cp $a $dist
 done
 
-# fire off mac and windows build machines
-if [ -z $SKIP_INSTALLERS ]; then
-  echo "starting installer jobs"
-  start_installer_job $WIN_JENKINS $WIN_JENKINS_USER $WIN_JENKINS_KEY $tag
-  start_installer_job $MAC_JENKINS $MAC_JENKINS_USER $MAC_JENKINS_KEY $tag
-fi
+cp $artifacts/../../../doc/en/$usertarget/latex/manual.pdf $dist/geoserver-$tag-user-manual.pdf
 
 # git commit changes on the release branch
 pushd .. > /dev/null
@@ -381,6 +402,13 @@ init_git $git_user $git_email
 git add . 
 git commit -m "updating version numbers and release notes for $tag" .
 popd > /dev/null
+
+# fire off mac and windows build machines
+if [ -z $SKIP_INSTALLERS ]; then
+  echo "starting installer jobs"
+  start_installer_job $WIN_JENKINS $WIN_JENKINS_USER $WIN_JENKINS_KEY $tag
+  start_installer_job $MAC_JENKINS $MAC_JENKINS_USER $MAC_JENKINS_KEY $tag
+fi
 
 popd > /dev/null
